requirements.txt
beautifulsoup4>=4.12,<5
PyYAML>=6,<7
requests>=2.31,<3

langchain-text-splitters>=0.3,<1

sentence-transformers>=3,<4
qdrant-client>=1.10,<2

langgraph>=0.2,<1
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
streamlit>=1.38,<2

pytest>=8,<10
